Changed data type of TotalCost

+++ src/main/java/com/group9/carrentalbackend/models/Rental.java
@@ -22,5 +22,5 @@ public class Rental {
     private Customer customer;
     private Date StartDate;
     private Date EndDate;
-    private Long TotalCost;
+    private Double TotalCost;
 }
